move public code to lib.rs
